move wasm module into legacy
