+++ src/legacy/wasm.rs
